CU012: PANTALLA DE INVENTARIO ARREGLADA, YA SE Y SE VEN LOS PEDIDOS PENDIENTES
