+++ src/main/java/co/ucentral/concesionario/persistencia/repositorios/InventarioRepositorio.java
@@ -2,7 +2,9 @@ package co.ucentral.concesionario.persistencia.repositorios;
 
 import co.ucentral.concesionario.persistencia.entidades.Inventario;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface InventarioRepositorio extends JpaRepository<Inventario, Long> {
     public Inventario findByVehiculoId(Long vehiculoId);
 }
