+++ src/main/resources/templates/pantallaAdministrador.html
@@ -158,7 +158,7 @@
 
 <script>
     function verInventario() {
-        mostrarModal('/pedidos/inventario');
+        mostrarModal('/inventario/verInventario');
     }
 
     function realizarPedido() {
